feat(dashboard): started working on permissions layout
